File added for Lab 5
Download the attached .zip archive, and open the HTML file in a code editor of your choice. Follow along with the numbered instructions in the HTML file, and preview the results as you progress by opening the same file in a browser. When you have finished, upload your files to your Web server and put the URL to your completed lab in the comments section in order for the lab to be graded.Lab 5 Start.zip